+++ lib/bootstrap.dart
@@ -1,0 +1,4 @@
+/// Provides [bootstrap] method to detect and add all root APIs
+library jaguar.bootstrap;
+
+export 'src/serve/bootstrap/bootstrap.dart';