+++ test/jaguar/exception/exception.g.dart
@@ -4,7 +4,6 @@ part of test.exception.exception;
 
 // **************************************************************************
 // Generator: ApiGenerator
-// Target: class ExampleApi
 // **************************************************************************
 
 class JaguarExampleApi implements RequestHandler {